Refactorize TreeHandler
